Backend: add Windows implementation of chooser classes
