+++ backend/CMakeLists.txt
@@ -4,9 +4,6 @@ set(SOURCES_PREFIX src)
 set(SOURCES
     ${SOURCES_PREFIX}/backend.cpp
     ${SOURCES_PREFIX}/keyinput.cpp
-    ${SOURCES_PREFIX}/choosepositionpushbutton.cpp
-    ${SOURCES_PREFIX}/choosewindowpushbutton.cpp
-    ${SOURCES_PREFIX}/targetwindow.cpp
     ${SOURCES_PREFIX}/windowhandle.cpp
     $<$<PLATFORM_ID:Windows>:${SOURCES_PREFIX}/backend-windows.cpp>
     $<$<PLATFORM_ID:Linux>:${SOURCES_PREFIX}/system-unix.cpp>
@@ -27,6 +24,9 @@ set(SOURCES
     $<$<PLATFORM_ID:Linux>:${SOURCES_PREFIX}/positionchooser-x11.cpp>
     $<$<PLATFORM_ID:Linux>:${SOURCES_PREFIX}/areachooser-x11.cpp>
     $<$<PLATFORM_ID:Linux>:${SOURCES_PREFIX}/windowchooser-x11.cpp>
+    $<$<PLATFORM_ID:Windows>:${SOURCES_PREFIX}/positionchooser-windows.cpp>
+    $<$<PLATFORM_ID:Windows>:${SOURCES_PREFIX}/areachooser-windows.cpp>
+    $<$<PLATFORM_ID:Windows>:${SOURCES_PREFIX}/windowchooser-windows.cpp>
 )
 
 set(HEADERS_PREFIX include/${PROJECT})
@@ -42,9 +42,6 @@ set(HEADERS
     ${HEADERS_PREFIX}/system.hpp
     ${HEADERS_PREFIX}/scopeguard.hpp
     ${HEADERS_PREFIX}/3rdparty/scope_guard/scope_guard.hpp
-    ${HEADERS_PREFIX}/choosepositionpushbutton.hpp
-    ${HEADERS_PREFIX}/choosewindowpushbutton.hpp
-    ${HEADERS_PREFIX}/targetwindow.hpp
     ${HEADERS_PREFIX}/windowhandle.hpp
     $<$<PLATFORM_ID:Windows>:${HEADERS_PREFIX}/backend-windows.hpp>
     $<$<PLATFORM_ID:Linux>:${HEADERS_PREFIX}/system-unix.hpp>
@@ -61,6 +58,9 @@ set(HEADERS
     $<$<PLATFORM_ID:Linux>:${HEADERS_PREFIX}/positionchooser-x11.hpp>
     $<$<PLATFORM_ID:Linux>:${HEADERS_PREFIX}/areachooser-x11.hpp>
     $<$<PLATFORM_ID:Linux>:${HEADERS_PREFIX}/windowchooser-x11.hpp>
+    $<$<PLATFORM_ID:Windows>:${HEADERS_PREFIX}/positionchooser-windows.hpp>
+    $<$<PLATFORM_ID:Windows>:${HEADERS_PREFIX}/areachooser-windows.hpp>
+    $<$<PLATFORM_ID:Windows>:${HEADERS_PREFIX}/windowchooser-windows.hpp>
 )
 
 add_library(${PROJECT} SHARED
