+++ backend/test/src/mainwindow.hpp
@@ -25,23 +25,13 @@ protected:
     void keyPressEvent(QKeyEvent *event) override;
 
 private slots:
-    void positionChooseStarted();
-    void positionChosen(QPointF position);
-    void foundValidWindow(const Backend::WindowHandle &handle);
-    void windowSearchEnded(const Backend::WindowHandle &handle);
-    void on_chooseRectanglePushButton_clicked();
-
     void on_chooseRectangleGnomePushButton_clicked();
-
-    void on_choosePositionNewPushButton_clicked();
-
     void on_chooseAreaPushButton_clicked();
-
-    void on_chooseWindowNewPushButton_clicked();
+    void on_chooseWindowNewPushButton_pressed();
+    void on_choosePositionNewPushButton_pressed();
 
 private:
     Ui::MainWindow *ui;
-    Backend::TargetWindow *mTargetWindow{};
     bool mDisableEscape{false};
 };
