+++ backend/test/src/mainwindow.ui
@@ -16,97 +16,55 @@
   <widget class="QWidget" name="centralwidget">
    <layout class="QFormLayout" name="formLayout">
     <item row="0" column="0">
-     <widget class="Backend::ChoosePositionPushButton" name="choosePositionPushButton">
+     <widget class="QPushButton" name="choosePositionNewPushButton">
       <property name="text">
        <string>Choose position</string>
       </property>
      </widget>
     </item>
     <item row="0" column="1">
-     <widget class="QLabel" name="positionLabel">
-      <property name="text">
-       <string/>
-      </property>
-     </widget>
-    </item>
-    <item row="5" column="0">
-     <widget class="Backend::ChooseWindowPushButton" name="chooseWindowPushButton">
-      <property name="text">
-       <string>Choose window</string>
-      </property>
-     </widget>
-    </item>
-    <item row="2" column="0">
-     <widget class="QPushButton" name="chooseRectanglePushButton">
-      <property name="text">
-       <string>Choose rectangle</string>
-      </property>
-     </widget>
-    </item>
-    <item row="2" column="1">
-     <widget class="QLabel" name="rectangleLabel">
-      <property name="text">
-       <string/>
-      </property>
-     </widget>
-    </item>
-    <item row="5" column="1">
-     <widget class="QLabel" name="windowLabel">
-      <property name="text">
-       <string/>
-      </property>
-     </widget>
-    </item>
-    <item row="4" column="0">
-     <widget class="QPushButton" name="chooseRectangleGnomePushButton">
-      <property name="text">
-       <string>Choose rectangle (gnome screenshot)</string>
-      </property>
-     </widget>
-    </item>
-    <item row="4" column="1">
-     <widget class="QLabel" name="rectangleGnomeLabel">
-      <property name="text">
-       <string/>
-      </property>
-     </widget>
-    </item>
-    <item row="1" column="0">
-     <widget class="QPushButton" name="choosePositionNewPushButton">
-      <property name="text">
-       <string>Choose position (new)</string>
-      </property>
-     </widget>
-    </item>
-    <item row="1" column="1">
      <widget class="QLabel" name="choosePositionNewLabel">
       <property name="text">
        <string/>
       </property>
      </widget>
     </item>
-    <item row="3" column="0">
+    <item row="1" column="0">
      <widget class="QPushButton" name="chooseAreaPushButton">
       <property name="text">
        <string>Choose area</string>
       </property>
      </widget>
     </item>
-    <item row="3" column="1">
+    <item row="1" column="1">
      <widget class="QLabel" name="chooseAreaLabel">
       <property name="text">
        <string/>
       </property>
      </widget>
     </item>
-    <item row="6" column="0">
-     <widget class="QPushButton" name="chooseWindowNewPushButton">
+    <item row="2" column="0">
+     <widget class="QPushButton" name="chooseRectangleGnomePushButton">
       <property name="text">
-       <string>Choose window (new)</string>
+       <string>Choose area (gnome screenshot)</string>
       </property>
      </widget>
     </item>
-    <item row="6" column="1">
+    <item row="2" column="1">
+     <widget class="QLabel" name="rectangleGnomeLabel">
+      <property name="text">
+       <string/>
+      </property>
+     </widget>
+    </item>
+    <item row="3" column="0">
+     <widget class="QPushButton" name="chooseWindowNewPushButton">
+      <property name="text">
+       <string>Choose window</string>
+      </property>
+     </widget>
+    </item>
+    <item row="3" column="1">
      <widget class="QLabel" name="chooseWindowNewLabel">
       <property name="text">
        <string/>
@@ -116,18 +74,6 @@
    </layout>
   </widget>
  </widget>
- <customwidgets>
-  <customwidget>
-   <class>Backend::ChoosePositionPushButton</class>
-   <extends>QPushButton</extends>
-   <header>backend/choosepositionpushbutton.hpp</header>
-  </customwidget>
-  <customwidget>
-   <class>Backend::ChooseWindowPushButton</class>
-   <extends>QPushButton</extends>
-   <header>backend/choosewindowpushbutton.hpp</header>
-  </customwidget>
- </customwidgets>
  <resources/>
  <connections/>
 </ui>
